+++ src/lib/index.ts
@@ -64,8 +64,9 @@ export { writeLod } from './writers/write-lod';
 export { writeGlb } from './writers/write-glb';
 export { writeVoxel } from './writers/write-voxel';
 export type { WriteVoxelOptions, VoxelMetadata } from './writers/write-voxel';
-export { fillExterior, simplifyForCapsule } from './voxel/nav-simplify';
-export type { NavSeed, NavSimplifyResult } from './voxel/nav-simplify';
+export { carveInterior } from './voxel/carve-interior';
+export { fillExterior } from './voxel/fill-exterior';
+export type { NavSeed, NavSimplifyResult } from './voxel/fill-exterior';
 export { marchingCubes } from './voxel/marching-cubes';
 export type { MarchingCubesMesh } from './voxel/marching-cubes';
 export { filterCluster } from './voxel/filter-cluster';
